// Frontend/src/app/interfaces/user.ts
export interface User {
modified_by_name: any;
modified_at: any;
created_by_name: any;
  id:number
  username: string;
  password:string;
  first_name: string;
  last_name: string;
  addresse: string;
  telephone: string;
  type_utlisateur: string;
  date_embauche: string;
  created_by: number;
  modified_by: number;
}